Refactor configs and stores
The current code was complicated by the need to keep up with
a config, a store, a path and a subpath for both the input &
the output. This combines all four parameters into a single
class so that only two instances need passing. Functionality
has been moved into instance methods where possible to reduce
leakage, though more refactoring could be applied.